+++ lib/src/common.g.dart
@@ -366,3 +366,27 @@ ContributorWeekStatistics _$ContributorWeekStatisticsFromJson(
   return ContributorWeekStatistics(
       json['w'] as int, json['a'] as int, json['d'] as int, json['c'] as int);
 }
+
+CodeSearchResults _$CodeSearchResultsFromJson(Map<String, dynamic> json) {
+  return CodeSearchResults()
+    ..totalCount = json['total_count'] as int
+    ..incompleteResults = json['incomplete_results'] as bool
+    ..items = json['items'] == null
+        ? null
+        : CodeSearchItem.fromJsonList(json['items'] as List);
+}
+
+CodeSearchItem _$CodeSearchItemFromJson(Map<String, dynamic> json) {
+  return CodeSearchItem()
+    ..name = json['name'] as String
+    ..path = json['path'] as String
+    ..sha = json['sha'] as String
+    ..url = json['url'] == null ? null : Uri.parse(json['url'] as String)
+    ..gitUrl =
+        json['git_url'] == null ? null : Uri.parse(json['git_url'] as String)
+    ..htmlUrl =
+        json['html_url'] == null ? null : Uri.parse(json['html_url'] as String)
+    ..repository = json['repository'] == null
+        ? null
+        : Repository.fromJSON(json['repository'] as Map<String, dynamic>);
+}