+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## v4.0.2
+## v4.1.0
 
 - Fix return type of `RepositoriesService.listContributors`.
 - Fix return type of `RepositoriesService.createRelease`.
@@ -9,6 +9,10 @@
   - Made associated classes `ContributorStatistics` and 
     `ContributorWeekStatistics` immutable. Since these classes are only meant as
     return values, we're not treating this as a breaking change.
+- Added `Stream<CodeSearchResults> github.search.code(...)` search API
+  - Made `CodeSearchResults` class to hold search results
+  - Made `CodeSearchItem` class to hold each search result item
+  - Added a code search example
 
 ## v4.0.1